requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
jinja2==3.1.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
aiosmtpd==1.4.4
email-validator==2.1.0
python-dotenv==1.0.0
dnspython==2.4.2
cryptography==41.0.7
